fix: replace faker avatar generation with default profile image URL resolution
